Fixed Position and relocated the code
